Change amount types from `Int64` to `Int32`

+++ src/haystack/discount.cr
@@ -1,6 +1,6 @@
 struct Haystack::Discount
   include Hapi::Resource
 
-  getter amount : Int64
+  getter amount : Int32
   getter type : Type
 end
