+++ src/haystack/invoice/line_item.cr
@@ -1,6 +1,6 @@
 struct Haystack::Invoice::LineItem
   include Hapi::Resource
 
-  getter amount : Int64
+  getter amount : Int32
   getter name : String
 end
