+++ src/haystack/tax.cr
@@ -1,6 +1,6 @@
 struct Haystack::Tax
   include Hapi::Resource
 
-  getter amount : Int64
+  getter amount : Int32
   getter name : String
 end
